Primary Changes: - Payment flow implemented - Initiate payment method initiates the payment - Confirm payment method confirms the payment - Also changed the status of the seats BOOKED, AVAILABLE accordingly if   payment fails or succeeds - Used Read Commited for initiating payment and Serializable for   confirming payment

+++ src/main/java/org/example/book_my_show/models/Payment.java
@@ -20,7 +20,7 @@ public class Payment extends BaseModel {
     private PaymentProvider paymentProvider;
 
     // Missed:
-    private int amount;
+    private Double amount;
 
     // 1 payment would belong to only 1 ticket strictly
     // 1 ticket can have many partial payments
